API/requirements.txt
Flask==2.2.3
numpy==1.24.2
Pillow==9.4.0
tensorflow<2.11.0
Flask
opencv-python
python-dotenv